refactored make_pv_data
